+++ src/scicloj/hanamicloth/v1/plotlycloth.clj
@@ -329,6 +329,7 @@
    :=inferred-group submap->group
    :=group :=inferred-group
    :=predictors [:=x]
+   :=design-matrix-spec submap->design-matrix-spec
    :=model-options {:model-type :fastmath/ols}
    :=histogram-nbins 10
    :=coordinates hc/RMV
@@ -417,8 +418,18 @@
 (def layer-text (mark-based-layer :text))
 
 
+(dag/defn-with-deps submap->design-matrix-spec
+  [=y =predictors]
+  [[=y]
+   (->> =predictors
+        (mapv (fn [k]
+                [k (list
+                    'identity
+                    (-> k name symbol))])))])
+
+
 (dag/defn-with-deps smooth-stat
-  [=dataset =x =y =predictors =group =model-options]
+  [=dataset =x =y =predictors =group =design-matrix-spec =model-options]
   (when-not (@=dataset =y)
     (throw (ex-info "missing =y column"
                     {:missing-column-name =y})))
@@ -434,24 +445,18 @@
                  (throw (ex-info "missing =group column"
                                  {:group =group
                                   :missing-column-name g}))))))
-  (let [design-matrix-spec [[=y]
-                            (->> =predictors
-                                 (mapv (fn [k]
-                                         [k (list
-                                             'identity
-                                             (-> k name symbol))])))]
-        predictions-fn (fn [ds]
+  (let [predictions-fn (fn [ds]
                          (let [model (-> ds
                                          (tc/drop-missing [=y])
                                          (#(apply design-matrix/create-design-matrix
                                                   %
-                                                  design-matrix-spec))
+                                                  =design-matrix-spec))
                                          (tc/select-columns (cons =y =predictors))
                                          (ml/train =model-options))]
                            (-> ds
                                (#(apply design-matrix/create-design-matrix
                                         %
-                                        design-matrix-spec))
+                                        =design-matrix-spec))
                                (ml/predict model)
                                =y)))]
     (if =group